requirements.txt
asgiref==3.8.1
autopep8==2.2.0
black==24.4.2
click==8.1.7
dj-database-url==0.5.0
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.11.1
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.1
